Add files via upload
A Lexical Analyzer in C is a program that reads source code and breaks it into meaningful tokens such as keywords, identifiers, operators, constants, and special characters. It processes the input line by line, skipping comments and whitespace, and identifies invalid tokens using error handling. This project is a fundamental part of compiler design, helping to convert code into a form suitable for syntax analysis.